Decouple setting elevator and wrist goal from moving to position
This change allows the default command to be move the wrist / elevator to the set target, rather than a predefined one.
The reason this is necessary is so that we can add in manual control, and when the manual control is done, hold the elevator
at whatever position it is at, rather than resetting back down to 0.

Now commands will set the target position, which is a `runOnce` so it doesn't block the defaultc command from running. To get the
wrist / elevator to move back to the stow position, we need to call `.stow()` in `onFalse`.

Manual controls can be added that simply increase / decrease the goal by some delta, and then this goal will be persisted until an 'automated' command's `onFalse` is triggered again.