Added Command Class for Note Tracking
A command class has been added for note tracking. In addition, the robot now limits how close it can get to the note before the motors stop (this is important because the camera angle on the robot will allow the note to be detected even when it is at the bumper).